move producer as global service
